pyproject.toml
[project]
name = "dep-audit-deepagent"
version = "0.1.0"
description = "Python dependency security auditor using LangChain Deep Agents — PTC/PTD token-efficient multi-agent architecture"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.4.0",
    "langchain-openai>=1.1.10",
    "structlog>=24.0.0",
    "rich>=14.3.3",
    "docker>=7.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.1",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.urls]
Repository = "https://github.com/shuknk8s/dep-audit-deepagent"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
